+++ src/main/java/roomescape/controller/ReservationTimeController.java
@@ -1,6 +1,7 @@
 package roomescape.controller;
 
 import java.util.List;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -18,6 +19,7 @@ import roomescape.service.ReservationTimeService;
 public class ReservationTimeController {
     private final ReservationTimeService reservationTimeService;
 
+    @Autowired
     public ReservationTimeController(ReservationTimeService reservationTimeService) {
         this.reservationTimeService = reservationTimeService;
     }
