+++ src/main/java/roomescape/repository/ReservationTimeJdbcRepository.java
@@ -1,0 +1,49 @@
+package roomescape.repository;
+
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import javax.sql.DataSource;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.core.RowMapper;
+import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
+import org.springframework.stereotype.Repository;
+import roomescape.model.ReservationTime;
+
+@Repository
+public class ReservationTimeJdbcRepository implements ReservationTimeRepository{
+    private static final RowMapper<ReservationTime> actorRowMapper = (resultSet, rowNum) -> {
+        return new ReservationTime(
+                resultSet.getLong("id"),
+                resultSet.getTime("start_at").toLocalTime()
+        );
+    };
+    private final JdbcTemplate jdbcTemplate;
+    private final SimpleJdbcInsert jdbcInsert;
+
+    @Autowired
+    public ReservationTimeJdbcRepository(JdbcTemplate jdbcTemplate, DataSource dataSource) {
+        this.jdbcTemplate = jdbcTemplate;
+        this.jdbcInsert = new SimpleJdbcInsert(dataSource)
+                .withTableName("RESERVATION_TIME")
+                .usingGeneratedKeyColumns("id");
+    }
+
+    public Long createTime(ReservationTime reservationTime) {
+        Map<String, String> params = new HashMap<>();
+        params.put("start_at", reservationTime.getStartAt().toString());
+
+        return jdbcInsert.executeAndReturnKey(params).longValue();
+    }
+
+    public List<ReservationTime> readTimes() {
+        String sql = "select id, start_at from reservation_time";
+        return jdbcTemplate.query(sql, actorRowMapper);
+    }
+
+    public int deleteTime(Long id) {
+        String sql = "delete from reservation_time where id = ?";
+        return jdbcTemplate.update(sql, id);
+    }
+}
