+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -1,10 +1,12 @@
 package roomescape.service;
 
 import java.util.List;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import roomescape.dto.ReservationTimeRequest;
 import roomescape.dto.ReservationTimeResponse;
 import roomescape.model.ReservationTime;
+import roomescape.repository.ReservationTimeJdbcRepository;
 import roomescape.repository.ReservationTimeRepository;
 
 @Service
@@ -12,6 +14,7 @@ public class ReservationTimeService {
     public static final int ONE_ROW = 1;
     private final ReservationTimeRepository reservationTimeRepository;
 
+    @Autowired
     public ReservationTimeService(ReservationTimeRepository reservationTimeRepository) {
         this.reservationTimeRepository = reservationTimeRepository;
     }
